requirements.txt
pynput
requests
pyaudio
pyperclip
pygetwindow
python-dotenv
